requirements.txt
Django==4.2.2
django-extensions
django-debug-toolbar
django-admin-rangefilter
django-import-export
django-storages
django-filter
djangorestframework
djangorestframework-xml
djangorestframework-yaml
djangorestframework-simplejwt
drf-yasg
pymemcache

ipython
boto3
Celery
requests

#flake
flake8
flake8-builtins
flake8-print

#pytest
pytest
pytest-django
pytest-cov
pytest-mock
model_bakery
